lead_gen_api/requirements.txt
anyio==4.9.0
apify_client==1.9.2
apify_shared==1.4.0
asgiref==3.8.1
certifi==2025.1.31
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.2.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
more-itertools==10.6.0
packaging==25.0
psycopg2==2.9.10
PyJWT==2.9.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
